backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
httpx==0.28.1
python-dotenv==1.1.0
mangum==0.17.0 